Split snapshot and audit tests
