+++ test/selenium/src/tests/test_audit_page.py
@@ -10,15 +10,13 @@ import pytest
 
 from lib import base
 from lib.constants import messages
-from lib.constants.element import Lhn
 from lib.entities.entities_factory import (
     AuditsFactory, AssessmentTemplatesFactory, AssessmentsFactory)
-from lib.page import dashboard
-from lib.service import webui_service, rest_service
+from lib.service import webui_service
 
 
 class TestAuditPage(base.Test):
-  """Tests for audit page."""
+  """Tests for audit functionality."""
 
   @pytest.fixture(scope="function")
   def create_and_clone_audit(
@@ -49,65 +47,6 @@ class TestAuditPage(base.Test):
         "issue": new_issue_rest[0], "asmt_tmpl": new_asmt_tmpl_rest[0],
         "expected_asmt_tmpl": expected_asmt_tmpl,
         "control": new_control_rest, "program": new_program_rest}
-
-  @pytest.fixture(scope="function")
-  def create_audit_and_update_original_control(
-      self, new_control_rest, new_program_rest, map_control_to_program_rest,
-      new_audit_rest, update_control_rest
-  ):
-    """Create Audit with snapshotable Control and update original Control under
-    Program via REST API.
-    Preconditions:
-    - Program, Control created via REST API.
-    - Control mapped to Program via REST API.
-    - Audit created under Program via REST API.
-    - Original Control updated via REST API.
-    """
-    return {
-        "audit": new_audit_rest[0], "program": new_program_rest,
-        "control": new_control_rest, "updated_control": update_control_rest}
-
-  @pytest.fixture(scope="function")
-  def create_audit_and_delete_original_control(
-      self, new_control_rest, new_program_rest, map_control_to_program_rest,
-      new_audit_rest, delete_control_rest
-  ):
-    """Create Audit with snapshotable Control and delete original Control under
-    Program via REST API.
-    Preconditions:
-    - Program, Control created via REST API.
-    - Control mapped to Program via REST API.
-    - Audit created under Program via REST API.
-    - Original Control deleted via REST API.
-    """
-    return {
-        "audit": new_audit_rest[0], "program": new_program_rest,
-        "control": new_control_rest}
-
-  @pytest.fixture(scope="function")
-  def create_audit_and_update_first_of_two_original_controls(
-      self, create_audit_and_update_original_control
-  ):
-    """Create Audit with snapshotable Control and update original Control under
-    Program via REST API. After that create second Control and map it to
-    Program via REST API.
-    Preconditions:
-    - Execution and return of fixture
-      'create_audit_and_update_original_control'.
-    - Second Control created via REST API.
-    - Second Control mapped to Program via REST API.
-    """
-    audit_with_one_control = create_audit_and_update_original_control
-    second_control = rest_service.ControlsService().create(count=1)[0]
-    rest_service.ObjectsOwnersService().create(objs=second_control)
-    rest_service.RelationshipsService().create(
-        src_obj=audit_with_one_control["program"], dest_objs=second_control)
-    return {
-        "audit": create_audit_and_update_original_control["audit"],
-        "program": audit_with_one_control["program"],
-        "control": audit_with_one_control["control"],
-        "updated_control": audit_with_one_control["updated_control"],
-        "second_control": second_control}
 
   @pytest.mark.smoke_tests
   def test_asmt_tmpl_creation(self, new_audit_rest, selenium):
@@ -175,153 +114,7 @@ class TestAuditPage(base.Test):
         messages.ERR_MSG_FORMAT.format(expected_asmts, actual_asmts))
 
   @pytest.mark.smoke_tests
-  def test_audit_contains_readonly_ver_of_control(
-      self, new_control_rest, new_program_rest, map_control_to_program_rest,
-      new_audit_rest, selenium
-  ):
-    """Check via UI that Audit contains read-only snapshotable Control.
-    Preconditions:
-    - Program, Control created via REST API.
-    - Control mapped to Program via REST API.
-    - Audit created under Program via REST API.
-    """
-    audit, _ = new_audit_rest
-    actual_controls_tab_count = (webui_service.ControlsService(selenium).
-                                 get_count_objs_from_tab(src_obj=audit))
-    assert len([new_control_rest]) == actual_controls_tab_count
-    is_control_editable = (
-        webui_service.ControlsService(selenium).is_obj_editable_via_info_panel(
-            src_obj=audit, obj=new_control_rest))
-    assert is_control_editable is False
-
-  @pytest.mark.smoke_tests
-  def test_audit_contains_snapshotable_control_after_updating_control(
-      self, create_audit_and_update_original_control, selenium
-  ):
-    """Check via UI that Audit contains snapshotable Control that does not
-    equal updated version Control without version updating Control.
-    Preconditions:
-    - Execution and return of fixture
-      'create_audit_and_update_original_control'.
-    """
-    audit_with_one_control = create_audit_and_update_original_control
-    audit = audit_with_one_control["audit"]
-    expected_control = audit_with_one_control["control"]
-    actual_controls_tab_count = (webui_service.ControlsService(selenium).
-                                 get_count_objs_from_tab(src_obj=audit))
-    assert len([expected_control]) == actual_controls_tab_count
-    actual_controls = (webui_service.ControlsService(selenium).
-                       get_list_objs_from_tree_view(src_obj=audit))
-    assert [expected_control] == actual_controls, (
-        messages.ERR_MSG_FORMAT.format([expected_control], actual_controls))
-
-  @pytest.mark.smoke_tests
-  def test_update_snapshotable_ver_after_updating_original_control(
-      self, create_audit_and_update_original_control, selenium
-  ):
-    """Check via UI that Audit contains snapshotable Control that up-to-date
-    with it actual state of original control after updating snapshot to latest
-    version.
-    Preconditions:
-    - Execution and return of fixture
-      'create_audit_and_update_original_control'.
-    """
-    audit_with_one_control = create_audit_and_update_original_control
-    audit = audit_with_one_control["audit"]
-    control = audit_with_one_control["control"]
-    expected_control = audit_with_one_control["updated_control"]
-    (webui_service.ControlsService(selenium).
-     update_obj_ver_via_info_panel(src_obj=audit, obj=control))
-    actual_controls_tab_count = (webui_service.ControlsService(selenium).
-                                 get_count_objs_from_tab(src_obj=audit))
-    assert len([expected_control]) == actual_controls_tab_count
-    actual_controls = (webui_service.ControlsService(selenium).
-                       get_list_objs_from_tree_view(src_obj=audit))
-    assert [expected_control] == actual_controls, (
-        messages.ERR_MSG_FORMAT.format([expected_control], actual_controls))
-
-  @pytest.mark.smoke_tests
-  @pytest.mark.skipif(True, reason="Issue in app GGRC-1196")
-  def test_audit_contains_snapshotable_control_after_deleting_original_control(
-      self, create_audit_and_delete_original_control, selenium
-  ):
-    """Check via UI that Audit contains snapshotable Control even after
-    deleting original control.
-    Snapshot does not have links to update version to latest state
-    and to view original Control under Program.
-    Preconditions:
-    - Execution and return of fixture
-      'create_audit_and_delete_original_control'.
-    """
-    audit_with_one_control = create_audit_and_delete_original_control
-    audit = audit_with_one_control["audit"]
-    expected_control = audit_with_one_control["control"]
-    actual_controls_tab_count = (webui_service.ControlsService(selenium).
-                                 get_count_objs_from_tab(src_obj=audit))
-    assert len([expected_control]) == actual_controls_tab_count
-    actual_controls = (webui_service.ControlsService(selenium).
-                       get_list_objs_from_tree_view(src_obj=audit))
-    assert [expected_control] == actual_controls, (
-        messages.ERR_MSG_FORMAT.format([expected_control], actual_controls))
-    is_control_updateable = (webui_service.ControlsService(selenium).
-                             is_obj_updateble_via_info_panel(
-        src_obj=audit, obj=expected_control))
-    is_control_openable = (webui_service.ControlsService(selenium).
-                           is_obj_page_exist_via_info_panel(
-        src_obj=audit, obj=expected_control))
-    assert is_control_updateable is False
-    assert is_control_openable is False
-
-  @pytest.mark.smoke_tests
-  def test_mapped_to_program_controls_does_not_added_to_existing_audit(
-      self, create_audit_and_update_first_of_two_original_controls, selenium
-  ):
-    """Check via UI that Audit contains snapshotable Control that equal to
-    original Control does not contain Control that was mapped to
-    Program after Audit creation.
-    Preconditions:
-    - Execution and return of fixture
-      'create_audit_and_update_first_of_two_original_controls'.
-    """
-    audit_with_two_controls = (
-        create_audit_and_update_first_of_two_original_controls)
-    audit = audit_with_two_controls["audit"]
-    expected_control = audit_with_two_controls["control"]
-    actual_controls_tab_count = (webui_service.ControlsService(selenium).
-                                 get_count_objs_from_tab(src_obj=audit))
-    assert len([expected_control]) == actual_controls_tab_count
-    actual_controls = (webui_service.ControlsService(selenium).
-                       get_list_objs_from_tree_view(src_obj=audit))
-    assert [expected_control] == actual_controls, (
-        messages.ERR_MSG_FORMAT.format([expected_control], actual_controls))
-
-  @pytest.mark.smoke_tests
-  def test_bulk_update_audit_objects_to_latest_ver(
-      self, create_audit_and_update_first_of_two_original_controls, selenium
-  ):
-    """Check via UI that Audit contains snapshotable Controls that up-to-date
-    with their actual states after bulk updated audit objects
-    to latest version.
-    Preconditions:
-    - Execution and return of fixture
-      'create_audit_and_update_first_of_two_original_controls'.
-    """
-    audit_with_two_controls = (
-        create_audit_and_update_first_of_two_original_controls)
-    audit = audit_with_two_controls["audit"]
-    expected_controls = [audit_with_two_controls["updated_control"],
-                         audit_with_two_controls["second_control"]]
-    (webui_service.AuditsService(selenium).
-     bulk_update_via_info_page(audit_obj=audit))
-    actual_controls_tab_count = (webui_service.ControlsService(selenium).
-                                 get_count_objs_from_tab(src_obj=audit))
-    assert len(expected_controls) == actual_controls_tab_count
-    actual_controls = (webui_service.ControlsService(selenium).
-                       get_list_objs_from_tree_view(src_obj=audit))
-    assert expected_controls == actual_controls, (
-        messages.ERR_MSG_FORMAT.format(expected_controls, actual_controls))
-
-  @pytest.mark.smoke_tests
+  @pytest.mark.cloning
   def test_cloned_audit_contains_new_attrs(self, create_and_clone_audit,
                                            selenium):
     """Check via UI that cloned Audit contains new predicted attributes.
@@ -334,6 +127,7 @@ class TestAuditPage(base.Test):
         messages.ERR_MSG_FORMAT.format(expected_audit, actual_audit))
 
   @pytest.mark.smoke_tests
+  @pytest.mark.cloning
   def test_non_clonable_objs_donot_move_to_cloned_audit(
       self, create_and_clone_audit, selenium
   ):
@@ -350,6 +144,7 @@ class TestAuditPage(base.Test):
     assert actual_asmts_tab_count == actual_issues_tab_count == 0
 
   @pytest.mark.smoke_tests
+  @pytest.mark.cloning
   def test_clonable_audit_related_objs_move_to_cloned_audit(
       self, create_and_clone_audit, selenium
   ):
@@ -367,6 +162,7 @@ class TestAuditPage(base.Test):
             [expected_asmt_tmpl], actual_asmt_tmpls))
 
   @pytest.mark.smoke_tests
+  @pytest.mark.cloning
   def test_clonable_not_audit_related_objs_move_to_cloned_audit(
       self, create_and_clone_audit, selenium
   ):
@@ -386,29 +182,3 @@ class TestAuditPage(base.Test):
     actual_objs = [actual_controls, actual_programs]
     assert expected_objs == actual_objs, (
         messages.ERR_MSG_FORMAT.format(expected_objs, actual_objs))
-
-  @pytest.mark.smoke_tests
-  @pytest.mark.parametrize("tab_name", [Lhn.ALL_OBJS, Lhn.MY_OBJS])
-  @pytest.mark.parametrize(
-      "version_of_ctrl, is_found",
-      [("control", False), ("updated_control", True)],
-      ids=["Snapshoted version is not found",
-           "Actual snapshotable control is presented"])
-  def test_search_snapshots_lhn(
-      self, create_audit_and_update_original_control, selenium,
-          version_of_ctrl, is_found, tab_name):
-    """Check via UI that LHN search not looking for snapshots."""
-    selenium.get(dashboard.Dashboard.URL)
-    lhn_menu = dashboard.Dashboard(selenium).open_lhn_menu()
-    lhn_menu.select_tab(tab_name)
-    control_title = (
-        create_audit_and_update_original_control[version_of_ctrl].title
-    )
-    lhn_menu.filter_query(control_title)
-    list_of_ctrls = (
-        lhn_menu.select_controls_or_objectives()
-        .select_controls().members_visible
-    )
-    assert (control_title in
-            [el.text for el in list_of_ctrls if len(list_of_ctrls) != 0]
-            ) == is_found
